renamed command from pricing to scrooge
