+++ src/ralph_scrooge/forms.py
@@ -18,7 +18,7 @@ from ralph_scrooge.models import (
     UsagePrice,
     UsageType,
 )
-from ralph_scrooge.management.commands.pricing_sync import (
+from ralph_scrooge.management.commands.scrooge_sync import (
     get_collect_plugins_names,
 )
 from ralph_scrooge.utils.common import ranges_overlap
