+++ src/ralph_scrooge/tests/management/commands/test_scrooge_sync.py
@@ -13,62 +13,62 @@ from django.test import TestCase
 from django.test.utils import override_settings
 
 from ralph.util import plugin
-from ralph_scrooge.management.commands import pricing_sync
+from ralph_scrooge.management.commands import scrooge_sync
 
 
 COLLECT_PLUGINS = ['asset', 'business_line', 'warehouse']
-COMMAND_NAME = pricing_sync.Command.__module__.split('.')[-1]
+COMMAND_NAME = scrooge_sync.Command.__module__.split('.')[-1]
 
 
-class TestPricingSync(TestCase):
+class TestScroogeSync(TestCase):
     def test_load_plugins(self):
-        pricing_sync._load_plugins()
-        self.assertTrue(pricing_sync.PLUGINS_LOADED)
+        scrooge_sync._load_plugins()
+        self.assertTrue(scrooge_sync.PLUGINS_LOADED)
         self.assertGreater(len(plugin.BY_NAME.get('scrooge', {})), 0)
 
     @override_settings(COLLECT_PLUGINS=COLLECT_PLUGINS)
     def test_get_collect_plugins_names(self):
         self.assertEquals(
-            pricing_sync.get_collect_plugins_names(),
+            scrooge_sync.get_collect_plugins_names(),
             COLLECT_PLUGINS
         )
 
-    @patch('ralph_scrooge.management.commands.pricing_sync.plugin.run')
+    @patch('ralph_scrooge.management.commands.scrooge_sync.plugin.run')
     def test_run_plugin(self, plugin_run_mock):
         plugin_run_mock.return_value = True, 'Everything OK'
         today = datetime.date.today()
-        pricing_sync._run_plugin('abc', today)
+        scrooge_sync._run_plugin('abc', today)
         plugin_run_mock.assert_called_with('scrooge', 'abc', today=today)
 
-    @patch('ralph_scrooge.management.commands.pricing_sync.plugin.run')
+    @patch('ralph_scrooge.management.commands.scrooge_sync.plugin.run')
     def test_run_plugin_failure(self, plugin_run_mock):
         plugin_run_mock.return_value = False, 'Something goes wrong'
         today = datetime.date.today()
-        with self.assertRaises(pricing_sync.PluginError):
-            pricing_sync._run_plugin('abc', today)
+        with self.assertRaises(scrooge_sync.PluginError):
+            scrooge_sync._run_plugin('abc', today)
         plugin_run_mock.assert_called_with('scrooge', 'abc', today=today)
 
-    @patch('ralph_scrooge.management.commands.pricing_sync.plugin.run')
+    @patch('ralph_scrooge.management.commands.scrooge_sync.plugin.run')
     def test_run_plugin_exception(self, plugin_run_mock):
         def side_effect(*args, **kwargs):
             raise Exception('Plugin error')
         plugin_run_mock.side_effect = side_effect
         today = datetime.date.today()
         with self.assertRaises(Exception):
-            pricing_sync._run_plugin('abc', today)
+            scrooge_sync._run_plugin('abc', today)
         plugin_run_mock.assert_called_with('scrooge', 'abc', today=today)
 
     @override_settings(COLLECT_PLUGINS=COLLECT_PLUGINS)
-    @patch('ralph_scrooge.management.commands.pricing_sync._run_plugin')
+    @patch('ralph_scrooge.management.commands.scrooge_sync._run_plugin')
     def test_run_plugins(self, run_plugin_mock):
         def side_effect(name, today):
             if name == 'business_line':
-                raise pricing_sync.PluginError()
+                raise scrooge_sync.PluginError()
             else:
                 return None
         run_plugin_mock.side_effect = side_effect
         today = datetime.date.today()
-        result = [r for r in pricing_sync.run_plugins(
+        result = [r for r in scrooge_sync.run_plugins(
             today,
             COLLECT_PLUGINS + ['abc', 'def']
         )]
@@ -87,7 +87,7 @@ class TestPricingSync(TestCase):
         ]))
 
     @override_settings(COLLECT_PLUGINS=COLLECT_PLUGINS)
-    @patch('ralph_scrooge.management.commands.pricing_sync.run_plugins')
+    @patch('ralph_scrooge.management.commands.scrooge_sync.run_plugins')
     def test_command(self, run_plugins_mock):
         run_plugins_mock.iter.return_value.__iter__.return_value = iter([
             ('business_line', False),
@@ -98,7 +98,7 @@ class TestPricingSync(TestCase):
         run_plugins_mock.assert_called_with(today, COLLECT_PLUGINS)
 
     @override_settings(COLLECT_PLUGINS=COLLECT_PLUGINS)
-    @patch('ralph_scrooge.management.commands.pricing_sync.run_plugins')
+    @patch('ralph_scrooge.management.commands.scrooge_sync.run_plugins')
     def test_command_today(self, run_plugins_mock):
         run_plugins_mock.iter.return_value.__iter__.return_value = iter([
             ('business_line', False),
@@ -109,7 +109,7 @@ class TestPricingSync(TestCase):
         run_plugins_mock.assert_called_with(today, COLLECT_PLUGINS)
 
     @override_settings(COLLECT_PLUGINS=COLLECT_PLUGINS)
-    @patch('ralph_scrooge.management.commands.pricing_sync.run_plugins')
+    @patch('ralph_scrooge.management.commands.scrooge_sync.run_plugins')
     def test_command_yesterday(self, run_plugins_mock):
         run_plugins_mock.iter.return_value.__iter__.return_value = iter([
             ('business_line', False),
@@ -120,7 +120,7 @@ class TestPricingSync(TestCase):
         run_plugins_mock.assert_called_with(yesterday, COLLECT_PLUGINS)
 
     @override_settings(COLLECT_PLUGINS=COLLECT_PLUGINS)
-    @patch('ralph_scrooge.management.commands.pricing_sync.run_plugins')
+    @patch('ralph_scrooge.management.commands.scrooge_sync.run_plugins')
     def test_command_run_only(self, run_plugins_mock):
         run_plugins_mock.iter.return_value.__iter__.return_value = iter([
             ('business_line', False),
